+++ src/client/components/Blockly/initContent/content.jsx
@@ -1,4 +1,4 @@
-const INITIAL_XML = '<xml xmlns="http://www.w3.org/1999/xhtml"><block type="text" x="70" y="30"><field name="TEXT"></field></block></xml>';
+const INITIAL_XML = '<xml xmlns="http://www.w3.org/1999/xhtml"></xml>';
 
 const INITIAL_TOOLBOX_XML = '<xml xmlns="http://www.w3.org/1999/xhtml" id="toolbox" style="display: none;">\n'
     + '  <category name="Logic" colour="#5C81A6">\n'
