+++ src/client/components/index.js
@@ -1,4 +1,4 @@
 export { default as SearchInput } from './SearchInput';
 export { default as StatusBullet } from './StatusBullet';
 export { default as RouteWithLayout } from './RouteWithLayout';
-export { default as Blockly } from './Blockly';
+export { default as MyBlockly } from './Blockly';
